Add Constraint to GraphSchema - move types in the types.py file
