+++ examples/customize/build_graph/components/pruners/graph_pruner.py
@@ -3,13 +3,11 @@
 import asyncio
 
 from neo4j_graphrag.experimental.components.graph_pruning import GraphPruning
-from neo4j_graphrag.experimental.components.schema import (
+from neo4j_graphrag.experimental.components.types import (
     GraphSchema,
     NodeType,
     PropertyType,
     RelationshipType,
-)
-from neo4j_graphrag.experimental.components.types import (
     Neo4jGraph,
     Neo4jNode,
     Neo4jRelationship,
