+++ examples/customize/build_graph/components/schema_builders/schema_from_text.py
@@ -14,6 +14,8 @@ from dotenv import load_dotenv
 
 from neo4j_graphrag.experimental.components.schema import (
     SchemaFromTextExtractor,
+)
+from neo4j_graphrag.experimental.components.types import (
     GraphSchema,
 )
 from neo4j_graphrag.llm import OpenAILLM
